+++ booking/src/main/java/com/drbsimon/booking/model/Reservation.java
@@ -1,4 +1,4 @@
-package com.drbsimon.booking.entity;
+package com.drbsimon.booking.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
